pyproject.toml
[project]
name = "haystack-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "haystack-ai>=2.8.0",
    "ipython>=8.30.0",
    "ollama-haystack>=2.1.1",
    "scikit-learn>=1.6.0",
    "streamlit-utils>=0.1.1",
    "streamlit>=1.41.1",
    "tqdm>=4.67.1",
    "pyre-check>=0.9.23",
    "hydra-core>=1.3.2",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
]
